+++ LedGroup.h
@@ -11,7 +11,7 @@ class LedGroup: public PrimitiveOutput {
     public:
     LedGroup(uint16_t x, uint16_t y, uint16_t colorb, uint16_t tcolor, int count, GFX4d *gfx);
     ~LedGroup();
-    void update(uint16_t value);
+    void update(int32_t value);
 };
 
 #endif
